# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Smart Home Energy Monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  currentPowerW: number;
  todayEnergyKwh: number;
  monthEnergyKwh: number;
  estimatedBillInr: number;
  voltageV: number;
  currentA: number;
  powerFactorKw: number;
  todayChangePercent: number;
  monthChangePercent: number;
  billChangePercent: number;
  systemStatus: string;
  lastUpdated: string;
  activeApplianceCount: number;
  totalApplianceCount: number;
}

export interface PowerHistoryPoint {
  time: string;
  powerW: number;
}

export interface LiveMetrics {
  timestamp: string;
  powerW: number;
  voltageV: number;
  currentA: number;
  frequency: number;
  history: PowerHistoryPoint[];
}

export interface EnergyDataPoint {
  label: string;
  energyKwh: number;
  date: string;
}

export interface ApplianceEnergyShare {
  name: string;
  percentage: number;
  energyKwh: number;
  color: string;
}

export type RecommendationSeverity = typeof RecommendationSeverity[keyof typeof RecommendationSeverity];


export const RecommendationSeverity = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface Recommendation {
  id: string;
  appliance: string;
  message: string;
  savingInr: number;
  severity: RecommendationSeverity;
}

export interface EnergyAnalytics {
  efficiencyScore: number;
  peakUsageHour: string;
  lowestUsageHour: string;
  avgDailyKwh: number;
  savingOpportunities: string[];
  recommendations: Recommendation[];
  weeklyTrend: EnergyDataPoint[];
  hourlyPattern: EnergyDataPoint[];
}

export interface Appliance {
  id: string;
  name: string;
  location: string;
  isOn: boolean;
  powerW: number;
  todayUsageKwh: number;
  monthlyCostInr: number;
  iconType: string;
  runningHoursToday: number;
}

export interface ApplianceToggleInput {
  isOn: boolean;
}

export interface TurnAllOffResult {
  success: boolean;
  message: string;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  high_power: 'high_power',
  unusual_usage: 'unusual_usage',
  bill_alert: 'bill_alert',
  appliance_alert: 'appliance_alert',
  info: 'info',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Alert {
  id: string;
  type: AlertType;
  title: string;
  message: string;
  timestamp: string;
  severity: AlertSeverity;
  isDismissed: boolean;
  /** @nullable */
  applianceName: string | null;
}

export interface PredictionInput {
  prevMonthKwh: number;
  currentMonthKwh: number;
  avgDailyKwh: number;
  applianceCount: number;
  /** @nullable */
  targetBudgetInr?: number | null;
}

export interface PredictionResult {
  predictedBillInr: number;
  predictedBillMinInr: number;
  predictedBillMaxInr: number;
  confidencePercent: number;
  expectedKwh: number;
  targetBudgetInr: number;
  budgetUsedPercent: number;
  willExceedBudget: boolean;
  tip: string;
  daysRemainingInMonth: number;
}

export interface UserSettings {
  userName: string;
  homeDescription: string;
  electricityProvider: string;
  tariffRatePerKwh: number;
  monthlyBudgetInr: number;
  highPowerThresholdW: number;
  notificationsEnabled: boolean;
  emailNotifications: boolean;
  smsNotifications: boolean;
  timezone: string;
}

export interface UserSettingsInput {
  userName?: string;
  homeDescription?: string;
  electricityProvider?: string;
  tariffRatePerKwh?: number;
  monthlyBudgetInr?: number;
  highPowerThresholdW?: number;
  notificationsEnabled?: boolean;
  emailNotifications?: boolean;
  smsNotifications?: boolean;
  timezone?: string;
}

export type GetEnergyConsumptionParams = {
/**
 * Time period for data
 */
period?: GetEnergyConsumptionPeriod;
};

export type GetEnergyConsumptionPeriod = typeof GetEnergyConsumptionPeriod[keyof typeof GetEnergyConsumptionPeriod];


export const GetEnergyConsumptionPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
  year: 'year',
} as const;