# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Smart Home Energy Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appliances/{id}/toggle:
    patch:
      operationId: toggleAppliance
      tags: [appliances]
      summary: Toggle appliance on/off
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplianceToggleInput"
      responses:
        "200":
          description: Updated appliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appliance"
...
  /alerts/{id}/dismiss:
    patch:
      operationId: dismissAlert
      tags: [alerts]
      summary: Dismiss an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  schemas:
...
    Appliance:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        isOn:
          type: boolean
        powerW:
          type: number
        todayUsageKwh:
          type: number
        monthlyCostInr:
          type: number
        iconType:
          type: string
        runningHoursToday:
          type: number
      required:
        - id
        - name
        - location
        - isOn
        - powerW
        - todayUsageKwh
        - monthlyCostInr
        - iconType
        - runningHoursToday
    ApplianceToggleInput:
      type: object
      properties:
        isOn:
          type: boolean
      required:
        - isOn
...
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [high_power, unusual_usage, bill_alert, appliance_alert, info]
        title:
          type: string
        message:
          type: string
        timestamp:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        isDismissed:
          type: boolean
        applianceName:
          type: ["string", "null"]
      required:
        - id
        - type
        - title
        - message
        - timestamp
        - severity
        - isDismissed
        - applianceName
...