# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Smart Home Energy Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';
import type { AlertType } from './alertType';

export interface Alert {
  id: string;
  type: AlertType;
  title: string;
  message: string;
  timestamp: string;
  severity: AlertSeverity;
  isDismissed: boolean;
  /** @nullable */
  applianceName: string | null;
}