# lib/api-zod/src/generated/types/appliance.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Smart Home Energy Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface Appliance {
  id: string;
  name: string;
  location: string;
  isOn: boolean;
  powerW: number;
  todayUsageKwh: number;
  monthlyCostInr: number;
  iconType: string;
  runningHoursToday: number;
}