# lib/api-zod/src/generated/types/applianceToggleInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AI Smart Home Energy Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplianceToggleInput {
  isOn: boolean;
}